Removing outdated page
Hello, I am currently working on updating the PlayReady.com website. My program manager brought to my attention that this page was outdated and needed to be removed. Since it is not on the PlayReady.com repository but here, I was wondering what the better way would be to delete this page from the Microsoft Learn website. Thank you for your help.